requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
